refactor: service, ui 계층을 module-api 로 이동
