+++ build.gradle
@@ -25,9 +25,6 @@ subprojects {
     }
 
     dependencies {
-        implementation 'org.springframework.boot:spring-boot-starter-actuator'
-        implementation 'org.springframework.boot:spring-boot-starter-jdbc'
-        implementation 'org.flywaydb:flyway-core'
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
     }
