+++ module-api/src/main/java/kitchenpos/ordertablegroup/application/OrderTableGroupService.java
@@ -1,5 +1,6 @@
 package kitchenpos.ordertablegroup.application;
 
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,19 +15,22 @@ import kitchenpos.ordertablegroup.dto.OrderTableGroupDto;
 public class OrderTableGroupService {
 	private final OrderTableGroupRepository orderTableGroupRepository;
 	private final OrderTableGroupValidator orderTableGroupValidator;
+	private final ApplicationEventPublisher applicationEventPublisher;
 
 	public OrderTableGroupService(
 		OrderTableGroupRepository orderTableGroupRepository,
-		OrderTableGroupValidator orderTableGroupValidator
+		OrderTableGroupValidator orderTableGroupValidator,
+		ApplicationEventPublisher applicationEventPublisher
 	) {
 		this.orderTableGroupRepository = orderTableGroupRepository;
 		this.orderTableGroupValidator = orderTableGroupValidator;
+		this.applicationEventPublisher = applicationEventPublisher;
 	}
 
 	@Transactional
 	public OrderTableGroupDto create(OrderTableGroupCreateRequest request) {
 		OrderTableGroup orderTableGroup = orderTableGroupRepository.save(OrderTableGroup.newInstance());
-		orderTableGroup.group(request.getOrderTableIds(), orderTableGroupValidator);
+		orderTableGroup.group(request.getOrderTableIds(), orderTableGroupValidator, applicationEventPublisher);
 		orderTableGroupRepository.save(orderTableGroup);
 		return OrderTableGroupDto.from(orderTableGroup);
 	}
@@ -35,7 +39,7 @@ public class OrderTableGroupService {
 	public void ungroup(Long id) {
 		OrderTableGroup orderTableGroup = orderTableGroupRepository.findById(id)
 			.orElseThrow(IllegalArgumentException::new);
-		orderTableGroup.ungroup(orderTableGroupValidator);
+		orderTableGroup.ungroup(orderTableGroupValidator, applicationEventPublisher);
 		orderTableGroupRepository.save(orderTableGroup);
 	}
 }
