+++ module-api/src/test/java/kitchenpos/menugroup/MenuGroupFixture.java
@@ -1,0 +1,17 @@
+package kitchenpos.menugroup;
+
+import kitchenpos.menugroup.dto.MenuGroupCreateRequest;
+
+public class MenuGroupFixture {
+	public static MenuGroupCreateRequest 추천_메뉴_그룹_요청() {
+		return new MenuGroupCreateRequest("추천메뉴");
+	}
+
+	public static MenuGroupCreateRequest 비추천_메뉴_그룹_요청() {
+		return new MenuGroupCreateRequest("비추천메뉴");
+	}
+
+	public static MenuGroupCreateRequest 이름없는_메뉴_그룹_요청() {
+		return new MenuGroupCreateRequest(null);
+	}
+}
