+++ module-domain/build.gradle
@@ -8,4 +8,5 @@ jar {
 
 dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.flywaydb:flyway-core'
 }
