+++ module-domain/src/main/java/kitchenpos/common/config/JpaAuditingConfiguration.java
@@ -1,0 +1,9 @@
+package kitchenpos.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfiguration {
+}
