+++ module-domain/src/main/resources/application.properties
