+++ module-domain/src/test/java/kitchenpos/order/OrderFixture.java
@@ -28,16 +28,4 @@ public class OrderFixture {
 			Collections.singletonList(OrderLineItemDto.from(후라이드후라이드_메뉴_주문_항목())),
 			new ValidOrderValidator());
 	}
-
-	public static OrderRequest 주문_요청(Long orderTableId, Long menuId, int quantity) {
-		return new OrderRequest(orderTableId, Collections.singletonList(new OrderLineItemDto(menuId, quantity)));
-	}
-
-	public static OrderRequest 주문_항목_없는_주문_요청(Long orderTableId) {
-		return new OrderRequest(orderTableId, Collections.emptyList());
-	}
-
-	public static OrderRequest 주문_상태_변경_요청(OrderStatus orderStatus) {
-		return new OrderRequest(orderStatus);
-	}
 }
