+++ module-domain/src/test/java/kitchenpos/ordertable/OrderTableFixture.java
@@ -2,7 +2,6 @@ package kitchenpos.ordertable;
 
 import kitchenpos.ordertable.domain.NumberOfGuests;
 import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.dto.OrderTableRequest;
 
 public class OrderTableFixture {
 	public static OrderTable 비어있지않은_주문_테이블_1번() {
@@ -27,25 +26,5 @@ public class OrderTableFixture {
 
 	public static OrderTable 그룹핑된_주문_테이블_6번() {
 		return OrderTable.of(6L, 1L, NumberOfGuests.from(0), true);
-	}
-
-	public static OrderTableRequest 빈_주문_테이블_요청() {
-		return new OrderTableRequest(4, true);
-	}
-
-	public static OrderTableRequest 비어있지않은_주문_테이블_요청() {
-		return new OrderTableRequest(4, false);
-	}
-
-	public static OrderTableRequest 비우기_요청() {
-		return new OrderTableRequest(true);
-	}
-
-	public static OrderTableRequest 채우기_요청() {
-		return new OrderTableRequest(false);
-	}
-
-	public static OrderTableRequest 손님_수_변경_요청(int numberOfGuests) {
-		return new OrderTableRequest(numberOfGuests);
 	}
 }
